NamedVectorizations.NV — Type

NV{T}(layout, vector)

The named vectorization core struct.

Fields

vector::Vector - the Vector representing the concatenation of the vectorized Array and Scalar elements.

layout::NamedTuple - it defines the position and dimension of the Arrays comprising the NV with respect to the vectorized representation. Each key-value pair is (array_name => (array_size, start, stop)) where start and stop bound the vector chunk containing the vectorized Array.

NamedVectorizations.vector — Method

vector(nv::NV)

Alias for getfield(nv, :vector).

NamedVectorizations.layout — Method

layout(nv::NV)

Alias for getfield(nv, :layout).